swh.lister.core: Remove abstractmethod
Some of the new listers like GNU and CRAN do not
follow the conventional way of making an HTTP
request, hence they do not need some of the methods
which are usually needed by in conventional HTTP
request.

But those method are marked abstractmethod in the
core making them necessary to be present. So  it is in
best to remove abstractmethod to increase the
readability of those listers.