maven: Simplify and improve maven repository listing process
Latest version of maven index exporter tool significantly reduces
processing time and disk space usage which makes it feasible to
directly execute the tool in a maven lister celery worker.

Update the maven lister implementation accordingly to the latest
changes in maven index exporter:
- run the index export tool directly in the lister (requires
  java JRE >= 17 in the execution environment)
- process produced NDJSON file by the index exporter to extract
  maven packages and their versions hosted in the repository

The updated implementation also significantly reduces the memory
consumption of the lister as maven packages info in the NDJSON
file are sorted by groupId and artifactId which allows to process
all versions of a package sequentially and remove the need to
store all package artifact versions in a global dictionary.

Precompiled maven index exporter java binaries and related scripts
are available in the /opt/maven-index-exporter directory of the
softwareheritage/maven-index-exporter docker image.