npm: Migrate lister to updated replication API
npm recently introduced breaking changes in its replication API,
used to list all available packages, which broke the lister.

As it is no longer possible to get package metadata directly from
the replication API endpoints, modify the lister to only exploits
the changes endpoint of the API. This endpoint allows to get the
new or modified packages incrementally. Instead of fetching the
real modification date for a package, use the current listing date
as last updated value for a package.

See https://github.com/orgs/community/discussions/152515

Related to #4715.