pattern: Rework batch sending of listed origins to scheduler
Previously the listed origins were sent by batch to the scheduler
in the send_origins method at each listed page. But some listers
implementations can return a small amount of origins per page which
results in a high amount of RPC calls in production that terribly
slow downs the listing process.

Prefer to create batch of origins directly in the run method of the
pattern lister to overcome that issue and ensure the same amount of
origins per batch (except the last one that can be smaller).