packagist: Ensure to only process packages modified since last listing
The lister is sending the If-Modified-Since request header with the date
of the last listing to check if a package was modified since then but
the response status code is 304 only when the If-Modified-Since value
matches exactly the Last-Modified one.

So almost all packages were considered updated when relisting, and their
metadata were fetched again while they are not needed.

To save some time and bandwidth when relisting, prefer to send a HEAD
request in first place to get the Last-Modified header value of a
package and compare it with the last listing date so not modified
packages can be filtered out.

Use a pool of threads to fetch package metadata to speedup listing.